requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ultralytics>=8.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pillow>=10.0.0